backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
httpx>=0.27.0
openai>=1.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
uvicorn>=0.30.0
websockets>=12.0
